+++ types/cyclonedx/tests/cyclonedx_v1_2-tests.ts
